==> 1a_target_sentiment_analysis.ipynb <==
import pandas as pd
import sys

from NewsSentiment import TargetSentimentClassifier

# Download the dataset and place under dataset/ folder

news = pd.read_csv('dataset/news.csv')
news_tweets = pd.read_csv('dataset/news_tweets.csv')
# user_resp = pd.read_csv('results/user_response.csv')

news.shape, news_tweets.shape

## Compuate target based sentiment

def get_sentiment_scores(news_tweets, ent):
    
    news_tweets_filtered = news_tweets[news_tweets.text.str.contains(ent)]
    proc_head = []
    senti = []
    j = 0
    tsc = TargetSentimentClassifier()
    for i, n in enumerate(news_tweets_filtered.text):
        # print(type(n))
        k = n.split(ent)
        # print(k)
        sn = dict()
        j+=1
        if(ent in n):
            # print(x)
            # proc_head.append(x)
            x = [k[0], ent, k[1]]
            sentiment = tsc.infer_from_text(k[0].strip('\n'), ent, k[1])
            # print(x, sentiment)
            # print(sentiment[0]['class_label'], sentiment[0]['class_prob'])
            sn[sentiment[0]['class_label']] = sentiment[0]['class_prob']
            sn[sentiment[1]['class_label']] = sentiment[1]['class_prob']
            sn[sentiment[2]['class_label']] = sentiment[2]['class_prob']
            senti.append(sn)
            print(j, end = "\r")
            # sys.exit()
        else:
            senti.append(None)
            
    pos_t_senti = [x['positive'] if x != None else None for x in senti]
    neg_t_senti = [x['negative'] if x != None else None for x in senti]
    neu_t_senti = [x['neutral'] if x != None else None for x in senti]
    
    news_tweets_filtered['biden_pos_new'] = pos_t_senti
    news_tweets_filtered['biden_neg_new'] = neg_t_senti
    news_tweets_filtered['biden_neu_new'] = neu_t_senti
            
    return news_tweets_filtered

news_tweets_filtered_B = get_sentiment_scores(news_tweets, 'Biden')
news_tweets_filtered_T = get_sentiment_scores(news_tweets, 'Trump')

news_tweets_senti = pd.concat((news_tweets_filtered_B, news_tweets_filtered_T), axis = 0)

news_tweets_filtered = news_tweets_filtered.drop_duplicates('tweet_id')

news_tweets_senti.to_csv('results/news_tweets.csv')
