==> 3b_user_response_emotion_analysis.ipynb <==
import pandas as pd
import numpy as np
import sys

import matplotlib.pyplot as plt
# from bertopic import BERTopic
pd.set_option('display.max_colwidth', None)
from scipy import stats

user_resp = pd.read_csv('results/user_response.csv')
news_tweets = pd.read_csv('results/news_tweets.csv')
# news_tweets = pd.read_csv('results/older_sentiment/news_tweets.csv')

user_resp.publication.value_counts()

user_resp.shape

user_resp['publish_date'] = pd.to_datetime(user_resp['created_at'])
user_resp['day'] = user_resp.publish_date.dt.day
user_resp['month'] = user_resp.publish_date.dt.month
user_resp['year'] = user_resp.publish_date.dt.year
user_resp['timestamp'] = user_resp.publish_date.dt.time

user_resp.shape

user_resp.conversation_id.nunique(), news_tweets.conversation_id.nunique()

R_user_resp = user_resp[((user_resp['publication'] == 'Breitbart News') | (user_resp['publication'] == 'Fox News'))]
L_user_resp = user_resp[((user_resp['publication'] == 'CNN') | (user_resp['publication'] == 'The Washington Post'))]
C_user_resp = user_resp[((user_resp['publication'] == 'Business Insider') | (user_resp['publication'] ==  'USA Today'))]

R_news_tweets = news_tweets[((news_tweets['publication'] == 'Breitbart News') | (news_tweets['publication'] == 'Fox News'))]
L_news_tweets = news_tweets[((news_tweets['publication'] == 'CNN') | (news_tweets['publication'] == 'The Washington Post'))]
C_news_tweets = news_tweets[((news_tweets['publication'] == 'Business Insider') | (news_tweets['publication'] ==  'USA Today'))]

R_news_tweets.shape, C_news_tweets.shape, L_news_tweets.shape

### Average number of user responses per news tweet

R_user_resp.shape[0]/R_user_resp.conversation_id.nunique(), L_user_resp.shape[0]/L_user_resp.conversation_id.nunique(), C_user_resp.shape[0]/C_user_resp.conversation_id.nunique()

cols = ['topics', 'like_count', 'quote_count', 'reply_count', 'retweet_count', 'anger', 'joy', 'optimism', 'sadness', 'trump_pos', 'trump_neg', 'trump_neu', 'biden_pos', 'biden_neg', 'biden_neu']

data = []
for top, grp in R_user_resp.groupby('topics'):
    num_resp = grp.shape[0]/grp.conversation_id.nunique()
    
    data.append((top, num_resp))
df_resp_pt_R = pd.DataFrame(data, columns = ['topics', 'resp_per_tweet'])

data = []
for top, grp in L_user_resp.groupby('topics'):
    num_resp = grp.shape[0]/grp.conversation_id.nunique()
    
    data.append((top, num_resp))
df_resp_pt_L = pd.DataFrame(data, columns = ['topics', 'resp_per_tweet'])

# for top in set(news_tweets.topics):
    
#     print(top, R_news_tweets[R_news_tweets.topics == top].Trump_flag.value_counts())

news_tweets[news_tweets.topics == 'capitol']

user_resp[user_resp.conversation_id.isin(news_tweets[news_tweets.topics == 'conspiracy_theory'].conversation_id)].subtopic.value_counts()

user_resp[user_resp.topics == 'fact-check'].subtopic.value_counts()

data = []
for top, grp in C_user_resp.groupby('topics'):
    num_resp = grp.shape[0]/grp.conversation_id.nunique()
    
    data.append((top, num_resp))
df_resp_pt_C = pd.DataFrame(data, columns = ['topics', 'resp_per_tweet'])

df_resp_pt_L.merge(df_resp_pt_C, on = 'topics').merge(df_resp_pt_R, on = 'topics').round()

C_user_resp.shape[0]/C_user_resp.conversation_id.nunique(), L_user_resp.shape[0]/L_user_resp.conversation_id.nunique(), R_user_resp.shape[0]/R_user_resp.conversation_id.nunique()

# R_user_resp[cols].describe()

# L_user_resp[cols].describe()

# C_user_resp[cols].describe()

cols = ['topics', 'like_count', 'quote_count', 'reply_count', 'retweet_count', 'anger', 'joy', 'optimism', 'sadness', 'trump_pos', 'trump_neg', 'trump_neu', 'biden_pos', 'biden_neg', 'biden_neu']

### Compare mean emotion values (L vs R vs C) across topics

R_user_resp_topic_mean = R_user_resp[cols].groupby('topics').mean()
L_user_resp_topic_mean = L_user_resp[cols].groupby('topics').mean()
C_user_resp_topic_mean = C_user_resp[cols].groupby('topics').mean()

R_user_resp_topic_mean.to_csv('results/user_resp_analysis/R_mean_emotions.csv')
L_user_resp_topic_mean.to_csv('results/user_resp_analysis/L_mean_emotions.csv')
C_user_resp_topic_mean.to_csv('results/user_resp_analysis/C_mean_emotions.csv')

#### Comparison based on Positive or negative mentions of Trump and Biden

def get_pos_neg_mentions(df, entity):
    
    entity_mentions = df[df[entity.title() + '_flag'] == True]
    pos_mean = entity_mentions[entity + '_pos'].mean()
    neu_mean = entity_mentions[entity + '_neu'].mean()
    neg_mean = entity_mentions[entity + '_neg'].mean()
    pos_entity_mentions = entity_mentions[((entity_mentions[entity + '_pos'] > pos_mean))]
    neg_entity_mentions = entity_mentions[((entity_mentions[entity + '_neg'] > neg_mean))]
    
    return pos_entity_mentions, neg_entity_mentions

T_pos_mentions_R,  T_neg_mentions_R = get_pos_neg_mentions(R_user_resp, "trump")
B_pos_mentions_R,  B_neg_mentions_R = get_pos_neg_mentions(R_user_resp, "biden")

T_pos_mentions_L,  T_neg_mentions_L = get_pos_neg_mentions(L_user_resp, "trump")
B_pos_mentions_L,  B_neg_mentions_L = get_pos_neg_mentions(L_user_resp, "biden")

# import seaborn as sns
# sns.displot(R_user_resp.trump_pos)

T_pos_mentions_R.shape, T_neg_mentions_R.shape, B_pos_mentions_R.shape, B_neg_mentions_R.shape

T_pos_mentions_L.shape, T_neg_mentions_L.shape, B_pos_mentions_L.shape, B_neg_mentions_L.shape

def get_mean_emotions_for_pos_neg_entity_mentions(df1, df2, ent, cols):
    
    return pd.concat((pd.DataFrame(df1[cols].mean(), columns = [ent + '_R']), df2[cols].mean()), axis = 1).rename(columns={0: ent + "_L"})
    # return pd.concat((pd.DataFrame(df1[cols].mean(), columns = [ent + '_pos']), df2[cols].mean()), axis = 1).rename(columns={0: ent + "_neg"})

_df1 = get_mean_emotions_for_pos_neg_entity_mentions(T_pos_mentions_R, T_pos_mentions_L, ent = 'Trump_pos', cols = ['anger', 'joy', 'optimism', 'sadness'])
_df2 = get_mean_emotions_for_pos_neg_entity_mentions(T_neg_mentions_R, T_neg_mentions_L, ent = 'Trump_neg', cols = ['anger', 'joy', 'optimism', 'sadness'])
_df3 = get_mean_emotions_for_pos_neg_entity_mentions(B_pos_mentions_R, B_pos_mentions_L, ent = 'Biden_pos', cols = ['anger', 'joy', 'optimism', 'sadness'])
_df4 = get_mean_emotions_for_pos_neg_entity_mentions(B_neg_mentions_R, B_neg_mentions_L, ent = 'Biden_neg', cols = ['anger', 'joy', 'optimism', 'sadness'])

# _df1 = get_mean_emotions_for_pos_neg_entity_mentions(T_pos_mentions_R, T_neg_mentions_R, ent = 'Trump_R', cols = ['anger', 'joy', 'optimism', 'sadness'])
# _df2 = get_mean_emotions_for_pos_neg_entity_mentions(T_pos_mentions_L, T_neg_mentions_L, ent = 'Trump_L', cols = ['anger', 'joy', 'optimism', 'sadness'])
# _df3 = get_mean_emotions_for_pos_neg_entity_mentions(B_pos_mentions_R, B_neg_mentions_R, ent = 'Biden_R', cols = ['anger', 'joy', 'optimism', 'sadness'])
# _df4 = get_mean_emotions_for_pos_neg_entity_mentions(B_pos_mentions_L, B_neg_mentions_L, ent = 'Biden_L', cols = ['anger', 'joy', 'optimism', 'sadness'])

comb_df = pd.concat((_df1, _df2, _df3, _df4), axis = 1)
comb_df

# comb_df.to_csv('results/user_resp_analysis/mean_emotions_for_pos_vs_neg_mentions.csv')

cols = ['anger', 'joy', 'optimism', 'sadness']

pd.concat((B_pos_mentions_L.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_B-L-pos', 'joy': 'joy_B-L-pos', 'optimism': 'optimism_B-L-pos', 'sadness': 'sadness_B-L-pos'}),
           B_neg_mentions_L.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_B-L-neg', 'joy': 'joy_B-L-neg', 'optimism': 'optimism_B-L-neg', 'sadness': 'sadness_B-L-neg'}),
           B_pos_mentions_R.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_B-R-pos', 'joy': 'joy_B-R-pos', 'optimism': 'optimism_B-R-pos', 'sadness': 'sadness_B-R-pos'}),
           B_neg_mentions_R.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_B-R-neg', 'joy': 'joy_B-R-neg', 'optimism': 'optimism_B-R-neg', 'sadness': 'sadness_B-R-neg'}),
           T_pos_mentions_L.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_T-L-pos', 'joy': 'joy_T-L-pos', 'optimism': 'optimism_T-L-pos', 'sadness': 'sadness_T-L-pos'}),
           T_neg_mentions_L.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_T-L-neg', 'joy': 'joy_T-L-neg', 'optimism': 'optimism_T-L-neg', 'sadness': 'sadness_T-L-neg'}),
           T_pos_mentions_R.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_T-R-pos', 'joy': 'joy_T-R-pos', 'optimism': 'optimism_T-R-pos', 'sadness': 'sadness_T-R-pos'}),
           T_neg_mentions_R.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_T-R-neg', 'joy': 'joy_T-R-neg', 'optimism': 'optimism_T-R-neg', 'sadness': 'sadness_T-R-neg'})), 
           axis = 1).to_csv('results/user_resp_analysis/mean_emotions_pos_neg_men_topic_wise.csv')

B_neg_mentions_R.groupby('topics').mean()[cols].rename(columns = {'anger': 'anger_B-R-neg', 'joy': 'joy_B-R-neg', 'optimism': 'optimism_B-R-neg', 'sadness': 'sadness_B-R-neg'})

comb_df

L_user_resp.shape

def get_normalized_dist(dist):
    
    return (dist - dist.min()) / (dist.max() - dist.min())

def get_normalized_emotion_scores(df):
    
    df['anger_norm'] = get_normalized_dist(df.anger)
    df['optimism_norm'] = get_normalized_dist(df.optimism)
    df['sadness_norm'] = get_normalized_dist(df.sadness)
    df['joy_norm'] = get_normalized_dist(df.joy)
    
    return df

# T_pos_mentions_R = get_normalized_emotion_scores(T_pos_mentions_R)
# B_neg_mentions_R = get_normalized_emotion_scores(B_neg_mentions_R)
# T_neg_mentions_R = get_normalized_emotion_scores(T_neg_mentions_R)
# B_pos_mentions_R = get_normalized_emotion_scores(B_pos_mentions_R)

# T_pos_mentions_L = get_normalized_emotion_scores(T_pos_mentions_L)
# B_neg_mentions_L = get_normalized_emotion_scores(B_neg_mentions_L)
# T_neg_mentions_L = get_normalized_emotion_scores(T_neg_mentions_L)
# B_pos_mentions_L = get_normalized_emotion_scores(B_pos_mentions_L)

# R_fav = pd.concat((T_pos_mentions_R, B_neg_mentions_R), axis = 0)
# R_unfav = pd.concat((T_neg_mentions_R, B_pos_mentions_R), axis = 0)
# R_fav.shape, R_unfav.shape

# L_fav = pd.concat((T_neg_mentions_L, B_pos_mentions_L), axis = 0)
# L_unfav = pd.concat((T_pos_mentions_L, B_neg_mentions_L), axis = 0)
# L_fav.shape, L_unfav.shape

def get_pos_and_neg_mentions(df, pub, ent):

    pos_mentions = df[((df['publication'] == pub) & (df[ent + '_pos'] > df[ent + '_pos'].mean()))]
    neg_mentions = df[((df['publication'] == pub) & (df[ent + '_neg'] > df[ent + '_neg'].mean()))]
    
    return pos_mentions, neg_mentions

T_pos_men_CNN, T_neg_men_CNN = get_pos_and_neg_mentions(user_resp, 'CNN', 'trump')
B_pos_men_CNN, B_neg_men_CNN = get_pos_and_neg_mentions(user_resp, 'CNN', 'biden')

T_pos_men_TWP, T_neg_men_TWP = get_pos_and_neg_mentions(user_resp, 'The Washington Post', 'trump')
B_pos_men_TWP, B_neg_men_TWP = get_pos_and_neg_mentions(user_resp, 'The Washington Post', 'biden')

T_pos_men_UST, T_neg_men_UST = get_pos_and_neg_mentions(user_resp, 'USA Today', 'trump')
B_pos_men_UST, B_neg_men_UST = get_pos_and_neg_mentions(user_resp, 'USA Today', 'biden')

T_pos_men_BI, T_neg_men_BI = get_pos_and_neg_mentions(user_resp, 'Business Insider', 'trump')
B_pos_men_BI, B_neg_men_BI = get_pos_and_neg_mentions(user_resp, 'Business Insider', 'biden')

T_pos_men_FN, T_neg_men_FN = get_pos_and_neg_mentions(user_resp, 'Fox News', 'trump')
B_pos_men_FN, B_neg_men_FN = get_pos_and_neg_mentions(user_resp, 'Fox News', 'biden')

T_pos_men_BN, T_neg_men_BN = get_pos_and_neg_mentions(user_resp, 'Breitbart News', 'trump')
B_pos_men_BN, B_neg_men_BN = get_pos_and_neg_mentions(user_resp, 'Breitbart News', 'biden')

R_fav = pd.concat((T_pos_men_FN, T_pos_men_BN, B_neg_men_FN, B_neg_men_BN), axis = 0)
R_unfav = pd.concat((T_neg_men_FN, T_neg_men_BN, B_pos_men_FN, B_pos_men_BN), axis = 0)
print(R_fav.shape, R_unfav.shape)

L_fav = pd.concat((B_pos_men_CNN, B_pos_men_TWP, T_neg_men_CNN, T_neg_men_TWP), axis = 0)
L_unfav = pd.concat((T_pos_men_CNN, T_pos_men_TWP, B_neg_men_CNN, B_neg_men_TWP,), axis = 0)
L_fav.shape, L_unfav.shape

cols = ['anger', 'optimism', 'sadness', 'joy']
pd.DataFrame(L_fav[cols].mean())

pd.DataFrame(L_unfav[cols].mean())

pd.DataFrame(R_fav[cols].mean())

pd.DataFrame(R_unfav[cols].mean())

def get_mean_emotions_for_pos_neg_entity_mentions(df1, df2, ent, cols):
    
    return pd.concat((pd.DataFrame(df1[cols].mean(), columns = [ent + '_R']), df2[cols].mean()), axis = 1).rename(columns={0: ent + "_L"})

_df1 = get_mean_emotions_for_pos_neg_entity_mentions(R_fav, R_unfav, ent = 'Right', cols = ['anger', 'joy', 'optimism', 'sadness'])
_df2 = get_mean_emotions_for_pos_neg_entity_mentions(L_fav, L_unfav, ent = 'Left', cols = ['anger', 'joy', 'optimism', 'sadness'])
# _df3 = get_mean_emotions_for_pos_neg_entity_mentions(T_pos_mentions_L, T_neg_mentions_L, ent = 'Trump_L', cols = ['anger', 'joy', 'optimism', 'sadness'])
# _df4 = get_mean_emotions_for_pos_neg_entity_mentions(B_pos_mentions_L, B_neg_mentions_L, ent = 'Biden_L', cols = ['anger', 'joy', 'optimism', 'sadness'])
comb_df = pd.concat((_df1, _df2), axis = 1)
comb_df

R_fav.columns

def get_mean_emotions_for_pos_neg_entity_mentions(df1, df2, ent, cols):
    
    return pd.concat((pd.DataFrame(df1[cols].mean(), columns = [ent + '_R']), df2[cols].mean()), axis = 1).rename(columns={0: ent + "_L"})

_df1 = get_mean_emotions_for_pos_neg_entity_mentions(R_fav, R_unfav, ent = 'Right', cols = ['anger_norm', 'optimism_norm', 'sadness_norm', 'joy_norm'])
_df2 = get_mean_emotions_for_pos_neg_entity_mentions(L_fav, L_unfav, ent = 'Left', cols = ['anger_norm', 'optimism_norm', 'sadness_norm', 'joy_norm'])
# _df3 = get_mean_emotions_for_pos_neg_entity_mentions(T_pos_mentions_L, T_neg_mentions_L, ent = 'Trump_L', cols = ['anger', 'joy', 'optimism', 'sadness'])
# _df4 = get_mean_emotions_for_pos_neg_entity_mentions(B_pos_mentions_L, B_neg_mentions_L, ent = 'Biden_L', cols = ['anger', 'joy', 'optimism', 'sadness'])
comb_df = pd.concat((_df1, _df2), axis = 1)
comb_df

cols = ['anger_norm', 'optimism_norm', 'sadness_norm', 'joy_norm']
R_fav[cols]

### Mean emotions across topics (L vs R)

df = pd.DataFrame()
cols = ['anger', 'joy', 'optimism', 'sadness']
df = pd.concat((df, L_user_resp[cols].describe().loc['mean']), axis = 1)
df = pd.concat((df, C_user_resp[cols].describe().loc['mean']), axis = 1)
df = pd.concat((df, R_user_resp[cols].describe().loc['mean']), axis = 1)
# df.T.to_csv('results/user_resp_analysis/mean_emotions.csv')
df.T

### Fav-vs-UnFav

# # df_L_T = L_user_resp[L_user_resp.Trump_flag == True]

# L_unfav1, L_fav1 = get_pos_neg_mentions(L_user_resp[L_user_resp.Trump_flag == True], 'trump')
# L_fav2, L_unfav2 = get_pos_neg_mentions(L_user_resp[L_user_resp.Biden_flag == True], 'biden')

# L_fav = pd.concat((L_fav1, L_fav2), axis = 0)
# L_unfav = pd.concat((L_unfav1, L_unfav2), axis = 0)

# # df_L_T = L_user_resp[L_user_resp.Trump_flag == True]

# R_fav1, R_unfav1 = get_pos_neg_mentions(R_user_resp[R_user_resp.Trump_flag == True], 'trump')
# R_unfav2, R_fav2 = get_pos_neg_mentions(R_user_resp[R_user_resp.Biden_flag == True], 'biden')

# R_fav = pd.concat((R_fav1, R_fav2), axis = 0)
# R_unfav = pd.concat((R_unfav1, R_unfav2), axis = 0)
