==> 4a_topic_modeling_news_tweets.ipynb <==
import pandas as pd
from nltk.tokenize import word_tokenize, sent_tokenize

import nltk
from nltk import word_tokenize, FreqDist
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
nltk.download
nltk.download('wordnet')
nltk.download('stopwords')
from nltk.tokenize import TweetTokenizer
from nltk.corpus import stopwords

# news = pd.read_csv('final_dataset/news_text_with_senti_scores.csv')

# Download and place dataset in the dataset/ folder
news_tweets = pd.read_csv('results/news_tweets.csv')
user_resp = pd.read_csv('results/user_response.csv')

news_tweets.columns, news_tweets.shape

news_tweets = news_tweets.drop_duplicates(subset = ['text'])

import re
news_tweets['hashtag'] = news_tweets.text.apply(lambda x: re.findall(r'#(\w+)', x))

import preprocessor as p
#forming a separate feature for cleaned tweets
clean_tweet = []
for i, v in enumerate(news_tweets['text']):
    # print(v)
    clean_tweet.append(p.clean(v))
    
news_tweets['clean_text'] = clean_tweet

def preprocess_data(data):
    
    #Removes Numbers
    data = data.astype(str).str.replace('\d+', '')
    lower_text = data.str.lower()
    lemmatizer = nltk.stem.WordNetLemmatizer()
    w_tokenizer =  TweetTokenizer()

    def lemmatize_text(text):
        return [(lemmatizer.lemmatize(w)) for w \
                       in w_tokenizer.tokenize((text))]
    
    def preprocess(txt):

        stop_words = set(stopwords.words('english'))
        exclude_words = ['donald', 'trump', 'joe', 'biden', 'president']
        return [w for w in txt if (not w.lower() in stop_words and not w.lower() in exclude_words)]
    
    def remove_punctuation(words):
        new_words = []
        # print(words)
        # sys.exit()
        for word in words:
          new_word = re.sub(r'[^\w\s]', '', (word))
          if new_word != '':
             new_words.append(new_word)
        # print(new_words)
        # sys.exit()
        
        return new_words
    
    words = lower_text.apply(lemmatize_text)
    words = words.apply(remove_punctuation)
    words = words.apply(preprocess)
    
    # print(words)
    # sys.exit()
    
    return pd.DataFrame(words)

proc_text = preprocess_data(news_tweets['clean_text'])
# proc_text = preprocess(proc_text.clean_text.to_list())

len(proc_text)

news_tweets['proc_text'] = proc_text

news_tweets = news_tweets.drop_duplicates(subset = ['proc_text'])

len(proc_tweets), news_tweets.shape

proc_tweets = [' '.join(map(str, l)) for l in news_tweets['proc_text']]
news_tweets['processed_text'] = proc_tweets

from bertopic import BERTopic
from umap import UMAP
from hdbscan import HDBSCAN
from sklearn.feature_extraction.text import CountVectorizer

seed_topic_list = [['drug', 'outbreak', 'flu', 'infection', 'contagious', 'treatment', 'prescription', 'covid', 'test', 'virus', 'ventilator', 'deaths', 'cases', 'pandemic', 
                    'epidemic', 'corona', 'coronavirus', 'covid19', 'patients', 'symptom'],
                   ['lockdown', 'shutdown', 'mask', 'distancing', 'masks'],
                   ['vaccine', 'vaccination', 'cure', 'moderna', 'pfizer', 'astrazeneca', 'j&j', 'dose'],
                   ['education', 'school', 'student', 'teacher', 'children', 'homeschool', 'schools', 'students', 'teachers'],
                   ['economy', 'industry', 'business', 'financial', 'finance', 'fiscal', 'economic', 'job', 'jobless', 'investing', 'investor', 'billion', 'gdp', 'debt', 
                    'liquidity', 'inflation', 'stimulus', 'bill', 'stocks', 'market', 'employment', 'unemployment', 'checks', 'cheques', 'recession', 'bull', 'bullish', 'bear',
                    'bearish', 'dow', 's&p', 'nasdaq', 'trade', 'trading', 'tax', 'loan', 'labor', 'buyback', 'selloff', 'wealth', 'wealthy', 'billionare', 'millionare'],
                   ['earth', 'green', 'pollution', 'ozone', 'deforestation', 'greenhouse', 'wildfire', 'climate', 'warming', 'temperature', 'flood', 'drought', 'glacier', 'environment', 'environmental', 'carbon', 'emission', 'gas', 'fracking'],
                   ['capitol', 'riot', 'siege', 'rioter', 'mob'],
                   ['voter', 'absentee', 'ballot', 'fraud', 'mailin', 'stolen', 'voting', 'election', 'black voters'],
                   ['immigration', 'immigrant', 'refugee', 'border', 'wall', 'migration', 'h1b', 'visa'],
                   ['blm', 'floyd', 'police', 'brutality', 'defund', 'protest', 'protesters', 'officer', 'black lives matter', 'injustice', 'racism', 'racial', 'supremacist'],
                   ['abortion', 'wade', 'roe'],
                   ['supreme', 'court', 'coney', 'barret', 'packing', 'justice', 'judge'],
                   ['security', 'military', 'weapons', 'attack', 'defense', 'gun', 'shooting', 'pentagon'],
                   ['international', 'country', 'global', 'china', 'chinese', 'beijing', 'shanghai', 'iran', 'irani', 'iranian', 'tehran', 'afghanistan', 'afghan', 'afghani', 'afghanistani', 
                    'kabul', 'russia', 'russian', 'moscow', 'britain', 'british', 'brit', 'brexit', 'london', 'Korea', 'Korean', 'kim', 'venezuelan', 'venezuela', 'syrian', 'syria'
                    'world', 'worldwide']]

my_stopwords = frozenset(list(["rt","RT", "&", "amp", "&amp", "http","https", "http://", "https://", "fav", "FAV"]))
vectorizer_model = CountVectorizer(ngram_range=(1, 2), stop_words = my_stopwords, min_df=5)

# do the BERT topic modelling
umap_model = UMAP(n_neighbors=10, n_components=5, min_dist=0.0, metric='cosine', random_state=42)

min_clusters = round(len(proc_tweets) * 0.0017)
hdbscan_model = HDBSCAN(min_cluster_size= min_clusters, metric='euclidean', cluster_selection_method='eom', prediction_data=True, min_samples=5)

# sentence_model = SentenceTransformer("all-mpnet-base-v2")
# embeddings = sentence_model.encode(news_list)

#run the model
topic_model = BERTopic(seed_topic_list = seed_topic_list, nr_topics = 'auto', umap_model=umap_model, hdbscan_model=hdbscan_model, vectorizer_model=vectorizer_model, low_memory=True, calculate_probabilities=True)

topics, probs = topic_model.fit_transform(proc_tweets)

news_tweets['topic_id'] = topics

topic_model.get_topic_info()

# news_tweets[news_tweets.topic_id == 26][['text', 'conversation_id']].iloc[0:50]

topic_model.get_topic(46)

#### Merging similar subtopics and grouping based on similarity after manual inspection

topic_list = {'election': {'fraud': [0, 71], 'electoral': [15], 'pres_deb': [18], 'opinion': [19, 30, 33], 'biden_transition': [1], 'rally': [6], 'general': [45, 52]},
              'covid': {'general': [2], 'mask': [8], 'response': [2, 9], 'vaccine': [14], 'school-reopen': [49, 86], 'treatment': [11], 'positive': [29], 'advisory': [43]},
              'republican': {'general': [5], 'gop': [63], 'william_barr': [35], 'pence': [55]},
              'white_house': {'general': [12], 'staff': [78, 83], 'first_lady': [58]},
              'democrats': {'harris': [13, 76], 'obama': [20], 'general': [37, 46], 'sander': [31], 'palosi': [62]},
              'national': {'security': [23], 'state': [34], 'executive_order': [25], 'admininstration': [32, 64], 'federal': [65, 85]},
              'blm': {'protest': [3]},
              'capitol': {'riot': [17]},
              'supreme_court': {'general': [4, 79]},
              'economy': {'stimulus': [7, 28], 'tax': [36]},
              'international': {'china': [16]},
              'impeachment': {'general': [27]},
              'media':{'twitter': [21], 'live': [24], 'fox': [41], 'usa_today': [77], 'news':[81]},
              'newyork': {'general': [10]},
              'american': {'people': [38, 73, 88], 'america_first': [42]},
              'healthcare': {'general': [51]},
              'conspiracy_theory': {'general':[44],'fact-check':[72, 75], 'perspective': [26]},
              'other': {'general': [22, 39, 40, 47, 48, 50, 53, 54, 56, 57, 59, 60, 61, 66, 67, 68, 69, 70, 74, 80, 84]},
              'immigration': {'general': [82, 87]}
             }

R_news_tweets = news_tweets[((news_tweets['publication'] == 'Breitbart News') | (news_tweets['publication'] == 'Fox News'))]
L_news_tweets = news_tweets[((news_tweets['publication'] == 'CNN') | (news_tweets['publication'] == 'The Washington Post'))]
C_news_tweets = news_tweets[((news_tweets['publication'] == 'Business Insider') | (news_tweets['publication'] ==  'USA Today'))]

# for top in set(news_tweets.topics):
    
#     print("- " * 50)
#     print(top)
#     print(R_news_tweets[R_news_tweets.topics == top][['like_count', 'quote_count', 'reply_count', 'retweet_count']].mean())
#     print(L_news_tweets[L_news_tweets.topics == top][['like_count', 'quote_count', 'reply_count', 'retweet_count']].mean())
#     print("- " * 50)
    

L_news_tweets[news_tweets.topics == 'conspiracy_theory'][['like_count', 'quote_count', 'reply_count', 'retweet_count']].mean()

pd.set_option('display.max_colwidth', None)
news_tweets[news_tweets.topics == 'conspiracy_theory'][['subtopic', 'text']]


# Uncomment to save model

# import pickle

# topic_model.save("bertopic/topic_model_news_tweets")
# pickle.dump(topics, open( "bertopic/topics_news_tweets.pickle", "wb" ) )
# # assert topics == new_topics

### Load saved model

from bertopic import BERTopic
import pickle
#Clear cache or restart kernel
my_model = BERTopic.load("../../bertopic/topic_model_news_tweets")

# topic_model.save("bertopic/topic_model_news_tweets")
# pickle.dump(topics, open( "bertopic/topics_news_tweets.pickle", "wb" ) )
# # assert topics == new_topics

my_model.get_topic_info()

# topic_list = {'covid': {'general': [16, 22], 'testing': [0], 'analysis': [8], 'mask': [12], 'cases': [15], 'fauci': [45], 'school-reopen': [56]},
#             'election': {'general': [9, 27, 28, 38, 40, 54, 63], 'fraud': [1], 'opinion': [4, 29, 69, 89], 'georgia': [25]},
#             'white_house': {'general': [2], 'staff': [90]},
#             'security': {'general': [3], 'intelligence': [85]},
#             'supreme_court': {'general': [5, 52]},
#             'blm': {'general': [6]},
#             'capitol': {'general': [7], 'impeachment': [68]},
#             'economy': {'stimulus': [10], 'tax': [51], 'stock': [81]},
#             'international': {'china': [14]},
#             'pres_debate': {'general': [17]},
#             'democarts':{'general': [19], 'kamala': [23, 59, 79], 'obama': [84], 'sanders': [32], 'biden': [39, 50, 66], 'pelosi': [75], 'primary': [60]},
#             'republican': {'general': [20], 'pence': [58], 'trump': [67], 'first_lady': [71]},
#             'others': {'general': [11, 13, 21, 24, 26, 30, 31, 33, 35, 36, 37, 43, 44, 46, 47, 48, 49, 53, 57, 61, 64, 70, 72, 73, 76, 77, 78, 82, 87, 88, 91, 92]},
#             'fact-check': {'general': [55], 'misleading': [74]},
#             'healthcare': {'general': [62]},
#             'immigration': {'border': [83]},
#             'social_media': {'general': [34, 86]},
#             'media': {'general': [18, 80, 93]},
#             'america': {'general': [41, 42, 65]}}

# pd.options.display.max_colwidth = 2000
# news_tweets[news_tweets['topic_ids'] == 41].text

# cls_to_fine_top[46]

### topic Ids
topic_cat_dic = dict([(j, i) for i, j in zip([key for key in topic_list], [x for x in range(len(topic_list))])])

### topic to cluster id mapping
topic_cls_map = dict()
for key in topic_list:
    cls_ids = []
    for k in topic_list[key]:
        cls_ids.extend(topic_list[key][k])
    topic_cls_map[key] =  cls_ids
    
cls_to_fine_top = dict()
for top in topic_list:
    for t in topic_list[top]:
        for k in topic_list[top][t]:
            cls_to_fine_top[k] = top + "-" + t
            
tops, fine_top = [], []
for t in topics:
    
    if(t == -1):
        tops.append('none')
        fine_top.append('none')
        continue
    for k in topic_cls_map:
        if(t in topic_cls_map[k]):
            tops.append(k)
    fine_top.append(cls_to_fine_top[t])

# cls_to_fine_top

len(tops), len(topics), len(fine_top)

news_tweets['topic_ids'] = topics
news_tweets['topics'] = tops
news_tweets['subtopic'] = fine_top

news_tweets.to_csv('dataset/news_tweets.csv')
