==> 5a_topic_analysis.ipynb <==
import pandas as pd
import sys

# Download and place dataset under the dataset/ folder

news = pd.read_csv('dataset/news.csv')
news_tweets = pd.read_csv('dataset/news_tweets.csv')
# user_resp = pd.read_csv('results/user_response.csv')

news.shape, news_tweets.shape

# news_filtered = news[news['topic_ids'] != -1]
# news_tweets_filtered = news_tweets[news_tweets['topic_ids'] != -1]

R_news = news[((news['publication'] == 'Breitbart News') | (news['publication'] == 'Fox News'))]
L_news = news[((news['publication'] == 'CNN') | (news['publication'] == 'The Washington Post'))]
C_news = news[((news['publication'] == 'Business Insider') | (news['publication'] ==  'USA Today'))]

R_news.shape, L_news.shape, C_news.shape

R_news_tweets = news_tweets[((news_tweets['publication'] == 'Breitbart News') | (news_tweets['publication'] == 'Fox News'))]
L_news_tweets = news_tweets[((news_tweets['publication'] == 'CNN') | (news_tweets['publication'] == 'The Washington Post'))]
C_news_tweets = news_tweets[((news_tweets['publication'] == 'Business Insider') | (news_tweets['publication'] ==  'USA Today'))]

R_news_tweets.shape, L_news_tweets.shape, C_news_tweets.shape

R_news[['trump_pos_new', 'trump_neg_new', 'trump_neu_new', 'biden_pos_new', 'biden_neg_new', 'biden_neu_new']].mean()

L_news[['trump_pos_new', 'trump_neg_new', 'trump_neu_new', 'biden_pos_new', 'biden_neg_new', 'biden_neu_new']].mean()

C_news[['trump_pos_new', 'trump_neg_new', 'trump_neu_new', 'biden_pos_new', 'biden_neg_new', 'biden_neu_new']].mean()

R_news_tweets.shape, L_news_tweets.shape, C_news_tweets.shape

news.subtopic.nunique(), news.topics.unique(), news_tweets.subtopic.nunique(), news_tweets.topics.unique()

### Topic-wise sentiment means

df1 = R_news.groupby('topics').mean()

df2 = L_news.groupby('topics').mean()

df3 = C_news.groupby('topics').mean()

df1.shape, df2.shape, df3.shape

df1 = df1[['trump_pos_new', 'trump_neg_new', 'trump_neu_new', 'biden_pos_new', 'biden_neg_new', 'biden_neu_new']]
df2 = df2[['trump_pos_new', 'trump_neg_new', 'trump_neu_new', 'biden_pos_new', 'biden_neg_new', 'biden_neu_new']]
df3 = df3[['trump_pos_new', 'trump_neg_new', 'trump_neu_new', 'biden_pos_new', 'biden_neg_new', 'biden_neu_new']]

df1['trump_fav_score'] = df1.trump_pos_new/df1.trump_neg_new
df2['trump_fav_score'] = df2.trump_pos_new/df2.trump_neg_new
df3['trump_fav_score'] = df3.trump_pos_new/df3.trump_neg_new

df1['biden_fav_score'] = df1.biden_pos_new/df1.biden_neg_new
df2['biden_fav_score'] = df2.biden_pos_new/df2.biden_neg_new
df3['biden_fav_score'] = df3.biden_pos_new/df3.biden_neg_new

df1.index.unique(), df2.index.unique(), df3.index.unique()

pd.concat((df1[['trump_fav_score', 'biden_fav_score']].rename(columns = {'trump_fav_score': 'trump_fav_score_R', 'biden_fav_score': 'biden_fav_score_R'}), 
           df2[['trump_fav_score', 'biden_fav_score']].rename(columns = {'trump_fav_score': 'trump_fav_score_L', 'biden_fav_score': 'biden_fav_score_L'}), 
           df3[['trump_fav_score', 'biden_fav_score']].rename(columns = {'trump_fav_score': 'trump_fav_score_C', 'biden_fav_score': 'biden_fav_score_C'})), axis = 1).to_csv('results/favorability_scores_news_headlines.csv')

# df1.reset_index(drop=True).merge(df2.reset_index(drop=True), left_index=True, right_index=True)
df1.to_csv('results/R_news_tweets_subtopic_based_mean_sentiments.csv')
df2.to_csv('results/L_news_tweets_subtopic_based_mean_sentiments.csv')

df2

# df1.reset_index(drop=True).merge(df2.reset_index(drop=True), left_index=True, right_index=True)
df1.to_csv('results/R_news_tweets_subtopic_based_mean_sentiments.csv')
df2.to_csv('results/L_news_tweets_subtopic_based_mean_sentiments.csv')

### Publisher-wise Trump vs Biden

news.publication.value_counts()

pubwise_news_mean = news.groupby('publication').mean()

pubwise_news_mean[['topic_ids', 'trump_pos_new', 'trump_neg_new', 'trump_neu_new', 'biden_pos_new', 'biden_neg_new', 'biden_neu_new']].to_csv('results/Trump-vs-Biden_news_tweets.csv')
